+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/i18n/I18n.kt
@@ -1316,6 +1316,11 @@ open class I18nView(
         DE - "Raum Id"
     }
 
+    open fun roomSettingsResetEncryption() = translate {
+        EN - "Reset outbound encryption"
+        DE - "Ausgehende Verschlüsselung zurücksetzen"
+    }
+
     open fun userSearchSearchPeople() = translate {
         EN - "search people"
         DE - "suche Personen"